--- cli-0/vite.config.ts
@@ -1,17 +1,0 @@
-import { defineConfig } from 'vite';
-import dts from 'vite-plugin-dts';
-
-export default defineConfig({
-    build: {
-        lib: {
-            entry: 'src/index.ts',
-            formats: ['es'],
-            fileName: 'index',
-        },
-        rollupOptions: {
-            external: ['commander'],
-        },
-        outDir: 'build',
-    },
-    plugins: [dts({ outDir: 'build' })],
-});
