+++ cli/src/setup/cra.ts
@@ -1,5 +1,5 @@
-import { CONFIG_OVERRIDES_FILE, CRA_COMMON_FILES, DEPENDENCY_NAME, JS_FILE_EXTENSION, ONLOOK_WEBPACK_PLUGIN, PACKAGE_JSON } from "./constants";
-import { exists, genASTParserOptionsByFileExtension, hasDependency } from "./utils";
+import { CONFIG_OVERRIDES_FILE, CRA_COMMON_FILES, DEPENDENCY_NAME, JS_FILE_EXTENSION, ONLOOK_WEBPACK_PLUGIN, PACKAGE_JSON } from "../constants";
+import { exists, genASTParserOptionsByFileExtension, hasDependency } from "../utils";
 
 import generate from '@babel/generator';
 import { parse } from '@babel/parser';
@@ -90,6 +90,7 @@ const ensureConfigOverrides = (): void => {
                     if (t.isMemberExpression(path.node.left) &&
                         t.isIdentifier(path.node.left.object, { name: 'module' }) &&
                         t.isIdentifier(path.node.left.property, { name: 'exports' })) {
+                        // @ts-ignore
                         path.node.right.arguments.push(
                             t.spreadElement(t.callExpression(t.identifier('addBabelPlugins'), [
                                 t.stringLiteral(ONLOOK_WEBPACK_PLUGIN)
